1.1.4a.3
I've created a switch in the Music_and_sound_design_platform patch that
turns on when the user runs the source code launcher. This will help us
detect whether the user is running source code or the application.

I've also added an interesting patch in the In Progress folder that will
allow us to add folders to the users Documents folder