Tweaks to Delay and Echo
Fixed the automation of the "Clear" buttons on both the Delay and Echo modules.

Also, the Delay module was renamed "Signal Delay" because all delay objects were being replaced with an abstraction of the Delay patch! Oh my! Talk about inefficient! I may want to reconsider the name of the module before 1.1.6