1.1.5.23
New Parameter Saving System is Complete! It has been implemented in all
of the templates as well as the Audio File Player. Once this is
implemented everywhere, we can begin to roll out the new preset system,
which has been designed to work with the new Parameter Saving System,
and can be found in the "_simplifier test" patch in the In-Progress
folder!