Project Settings tweaks and menu tweaks
The Project Settings screen is now properly titled, and appears along with the system board when a new project is created. I've also updated the MSDP menus, removing the full screen tabs, but making the Windows menu more organized.
I've also changed the default project settings so that the System Board opens NOT in full view, and set the default tab to Board Browser.