Finished Score Player and more
Removed the old project settings file
Tweaked the Score Player so that it works with it's new location in the System Board
Tweaked the current Project Settings 2 file so that scores load properly and to hide all metronome controls if metro is turned off.

IN PROGRESS:
Metronome controls still partially appear when user tabs over to them, even if metro is turned off. Has to do with show/hide of various processes on the System Board. Look into that in combination with Metronome toggler in Project Settings.