Drum Machine Default Kit and Save Issues
Lots of parameters aren't saving in the Drum Machine. They don't all have parameter numbers. Beat Source is now P50. Figure out the rest.

Default drumset for drum machine. In progress. Can point to the document tree in the saved preset, but need to put the files in the document tree.