load done BUT
load works BUT now there's a json issue. values are saved with four zeros after whatever the final decimal place is. This confuses line, which doesn't know what to do with that message. We need some kind of message formatting for json. Maybe put quotes around the value. In general I need to format the json values more