updates to master metro and chaos synth tweaks
I realized we need to store the current BPM, so there's now a value that gets used in the master metronome.

I'm also working on fixing several issues with timing in the chaos synth. here's what remains to tweak:

The Sliders do not always begin the first time that tab is opened. I need to switch out of the tab, then back in, then it works. This happened when I set the # of oscillators to 1. Doesn't happen if I keep the default #

In Keyboard mode, need to go back to the beginning of the pattern every time the on button starts

Tempo begins when a tab is clicked over to. This means that the tempo may be in-sync and may not be. Need to make sure that the beginning of the note aligns with the metro bang...