more progress
OK, there are now two metros. One for MSDP and one for MSDP4L. Test them both, and if they both work, make a piece of code to check version and chose metro solution accordingly. Combine it into a single abstraction, then implement it in place of the current solution.