tweaks and stripping
I've removed the javascript from the parameters, it was more trouble than it was worth. I also removed the ability to forward on to an optional second location, as this generates an error in the console if nothing is set. I also set the parameter I/O to go to the right inlet of the msdp.param objects, which allows the actual UI object to go to the left inlet. This makes less mess