play and note store removed
there can no longer be any auto-start on load of audio file playback and there is no pitch storage.

In the msdp.param.keyboard patch there is now an additional option as the second argument to determine whether or not to store pitch data with project. 0 for no, 1 for yes